+++ x/domain/client/rest/routes.go
@@ -15,18 +15,18 @@ import (
 // txRouteList clubs together all the transaction routes, which are the transactions
 // // that return the bytes to sign to send a request that modifies state to the domain module
 var txRoutesList = map[string]func(cliContext context.CLIContext) http.HandlerFunc{
-	"registerDomain":         registerDomainHandler,
-	"addAccountCertificates": addAccountCertificatesHandler,
-	"delAccountCertificates": delAccountCertificateHandler,
-	"deleteAccount":          deleteAccountHandler,
-	"deleteDomain":           deleteDomainHandler,
-	"registerAccount":        registerAccountHandler,
-	"renewAccount":           renewAccountHandler,
-	"renewDomain":            renewDomainHandler,
-	"replaceAccountTargets":  replaceAccountTargetsHandler,
-	"transferAccount":        transferAccountHandler,
-	"transferDomain":         transferDomainHandler,
-	"setAccountMetadata":     setAccountMetadataHandler,
+	"registerDomain":          registerDomainHandler,
+	"addAccountCertificates":  addAccountCertificatesHandler,
+	"delAccountCertificates":  delAccountCertificateHandler,
+	"deleteAccount":           deleteAccountHandler,
+	"deleteDomain":            deleteDomainHandler,
+	"registerAccount":         registerAccountHandler,
+	"renewAccount":            renewAccountHandler,
+	"renewDomain":             renewDomainHandler,
+	"replaceAccountResources": replaceAccountResourcesHandler,
+	"transferAccount":         transferAccountHandler,
+	"transferDomain":          transferDomainHandler,
+	"setAccountMetadata":      setAccountMetadataHandler,
 }
 
 // registerTxRoutes registers all the transaction routes to the router